Add asdf tool version file
The `.tool-versions` file is initially seeded with Poetry and Python.
The py version is on 3.7 as that is the latest version supported by AWS
Glue.

Poetry is at the latest version, which has support for py3.7. Others at
CTD have recommended it, and it seems like it should meet our needs.

Add a setup section to the README on how to install these dependencies.