Add Configuration Class to Library
This class will manage all of the differences between the 6 different
record json files that are being converted to parquet files. The initial
API has a Configuration Enum class that can be produced from a filename
and can generate a pyarrow schema that will be used to convert the json
to parquet.

This class was added to the library api and a testing stub has been
created that tests both methods.