Add Script to bach files by type up to total size
batch files dot py lives at the top level of the python directory, and
will take an input source of file information and compile it into a list
of Batch objects. These Batch objects contain a list of filenames and
are created to have a total filesize under a provided threshold.

batcher dot py was added to the py_gtfs_rt_ingestion library, and the
function to do the batching was added to init dot py, exposing it
explicitly as part of the modules API.