Add Error Handling and Logging
Add a file of custom library exceptions called error.py. It will contain
a base class for our custom errors and specific ones that inherit from
it. Use them where appropriate around the library.

Add logging to the library, replacing any print statements and ensure
that it is used at the info level in the two main scripts.