Add Batch.trigger_lambda functionality
The batch lambda function needs to be able to trigger the ingest lambda
for each batch that it creates. The method usesboto3, and an environment
variable that contains the ingest lambdas ARN. The actual call itself is
async and run inside of the s3_utils library. There are a handful of
logging statements and exceptions to make clear whats going on / wrong.

The batch lambda script has new flags that can be used to do a dry run
(not triggering the lambda) and printing out events as json formatted
strings.