Bug Fix  Move Lambda Types to gtfsrt ingest Module
The Lambda Types file was added in to provide type hints to the
parameters to lambda handler functions in both of our batch files and
ingest scripts. It was put at the top level next to both of these
scripts, but that causes problems when zipping up the project for
deployments. To avoid this, move it to the module itself so that it will
always be bundled.