Filter out s3 files based on last modified date
The Batching lambda is scheduled to run at the start of every hour.
However, due to start up times, the actual run time is a few minutes
after the start of the hour. This causes most batches to have files that
were created between 3 minutes past the previous hour up until 3 minutes
past the current hour. When the parquet files are generated from these
ingested files, we partition them, leaving us with two files. One is a
large file for all data created during the previous hour while the other
is quite small, containing only data created during the first three
minutes of the current hour.

Update `file_list_from_s3` utility to filter out files modified after
passed in datetime. This feature is used in batch files dot py to
prevent the files that have been uploaded since the start of the current
hour from being batched, which should help us avoid the splits happening
on partition.

Tests were updated to catch this behavior.

pyproject dot toml was also updated to allow log statements to come
through on failure, which is useful for debugging code both in the
module and in the tests.

https://app.asana.com/0/1189492770004753/1202506607268445/f