Filter Out GTFS RT Ingestion based on timestamp
The Batching lambda is scheduled to run at the start of every hour.
However, due to start up times, the actual run time is a few minutes
after the start of the hour. This causes most batches to have files that
were created between 3 minutes past the previous hour up until 3 minutes
past the current hour. When the parquet files are generated from these
ingested files, we partition them, leaving us with two files. One is a
large file for all data created during the previous hour while the other
is quite small, containing only data created during the first three
minutes of the current hour. Doing the filtering during batch would be
cleaner, but would require reading the gtfs_rt files themselves to the
their internal timestamps that we derive year, month, day, and hour
information from. Instead, filter out based on time during ingest.
Files that are too recent to neither archive nor error, so they will
remain in the incoming s3 bucket for the next hours pass.

* add `start_of_hour` member to `GtfsRtConverter` class that will be
  used to filter files that are too recent.
* in `gz_to_pyarrow` return None if the files internal timestamp is
  passed the start of the hour.
* add tests to check that `start_of_hour` has no minute, second, or
  microsecond information and that the filtering works as expected. a
  new vehicle positions file was added for this purpose.
* update pytest in pyproject dot toml to allow log statements to come
  through on failed tests, aiding debugging.

BUG FIX:
While working on this ticket I noticed that `GtfsRtConverter` was adding
files to the archive list when catching exceptions. That behavior was
patched and tests were updated to ensure everything was working as
expected.