HotFix: Handle Batch Args from events better
I changed how we create a BatchArgs out of an event in batch_files dot
py, thinking we could always parse them out of an event. Apparently
there are more fields being passed in on Lambda, and unpacking the event
to use as keyword value pairs in the Batch Args constructor causes a
TypeError from an unexpected keyword arg.

Revert back to old way (for now).