Update Database Schema and Calculations
After meeting with OPMI we are changing and simplifying the schema for
the database. Instead of an l0 table for vehicle positions and trip
updates and an l1 table that combines them, a single VehicleEvents table
will take its place. This table will eventually have runtimes and
headways as well.

* update the schema to simplify what we're doing here
* update the l0 vehicle position processing to generate events for new
  schema
* update the l0 trip updates to do the same
* remove dwell times, travel times, and headways computations