Change order of processing gtfs_rt files
We used to have two l0 load functions for gtfs_rt files, for trip
updates and vehicle positions respectively. In the new context of using
a VehicleEvents table, this meant that we would process vehicle position
data, upload it to the db, then pull it down again to add trip update
data and then upload it a second time. Now process both vehicle position
and trip update files, combine the datasets, and upload them in one go.

* tighten up the logic for calculating Vehicle Events from both Vehicle
  Positions and Trip Updates source files.
* improve comments and change from Int64 to int64 datatypes where
  appropriate
* add in process loggers to the vehicle events processing logic.
* add tests to check different cases, with utilities to monkey patch s3
  calls and check that logging is working correctly.
* update db manager to add files to the metadata table as a lost, which
  is used to seed files for tests.