Change order of processing gtfs_rt files
We used to have two l0 load functions for gtfs_rt files, for trip
updates and vehicle positions respectively. In the new context of using
a VehicleEvents table, this meant that we would process vehicle position
data, upload it to the db, then pull it down again to add trip update
data and then upload it a second time.

Now process both vehicle position and trip update files, combine the
datasets, and upload them in one go.