Vehicle Events Refactor (#91)
* Add Test Data for Performance Manager Tests

These tests use the same filestructure that ingestion creates for us on
s3 using prefixes. Its one days worth of trip updates and vehicle
position gtfs rt files. it also includes the schedule parquet files for
that day.

* Add l0 table creation tests

* update load parquet function in s3 utils dot py to be able to handle
  local or s3 file systems.
* add test fixtures that will add all of the test files to the metadtata
  table that will get picked up when loading l0 tables.
* add calls to load l0 table functions to first postgres test.

* Update Database Schema and Calculations

After meeting with OPMI we are changing and simplifying the schema for
the database. Instead of an l0 table for vehicle positions and trip
updates and an l1 table that combines them, a single VehicleEvents table
will take its place. This table will eventually have runtimes and
headways as well.

* update the schema to simplify what we're doing here
* update the l0 vehicle position processing to generate events for new
  schema
* update the l0 trip updates to do the same
* remove dwell times, travel times, and headways computations

* Change order of processing gtfs_rt files

We used to have two l0 load functions for gtfs_rt files, for trip
updates and vehicle positions respectively. In the new context of using
a VehicleEvents table, this meant that we would process vehicle position
data, upload it to the db, then pull it down again to add trip update
data and then upload it a second time. Now process both vehicle position
and trip update files, combine the datasets, and upload them in one go.

* tighten up the logic for calculating Vehicle Events from both Vehicle
  Positions and Trip Updates source files.
* improve comments and change from Int64 to int64 datatypes where
  appropriate
* add in process loggers to the vehicle events processing logic.
* add tests to check different cases, with utilities to monkey patch s3
  calls and check that logging is working correctly.
* update db manager to add files to the metadata table as a lost, which
  is used to seed files for tests.

* Small Table Updates post Events Refactor

Add Trip Hash to Vehicle Events Table. It is unique to each trip taken.
It should be useful for computing metrics later and for queries our
users want to run.

Update the Metadata Table name in ingestion so that it matches the new
name in performance manager.