Better org of pipeline and util files
Instead of having the scripts that startup our pipelines in a `startup`
directory, put them in the directories with all the code they use to
run. (so startup/start_ingestion -> ingestion/pipeline). this way, if we
change the starup file, it'll be caught by our bash script in the
determine_tests action. This required some imports to change, as well as
the script commands in the toml file.

Then move the two utility files into a new subdirectory called
runtime_utils. i left import_env with the same filename, though i
suspect this might turn into an argparse utils file. i renamed
logging_utils as process_logger, since thats a bit more accurate to what
the file contains. All import statements were adjusted accordingly. I
then added a line to the bash script that will check if the runtime
utils are changed, forcing tests of ingestion and performance manager if
so. (we should probably also make tests on these utils).