Update github workflows for more deploy options
We want to be able to deploy to our dev, staging, and production
environments. To do so, add in two new workflows for prod and staging.

Now is also a sensible time for us to change our behavior on merges to
main to run CI and update our staging environment rather than our dev
one. This will leave our dev one locked until we manually deploy using
the github UI.

For now, prod is also going to be deployed via manual clicks on the
github UI until we figure out how we want to control versioning via
tags.