Add in API Elixir Application (#112)
* Add in API Elixir Application

This application was created using the Phoenix framework. It currently
is configured to connect to a local running database we create via
docker compose. The app is able to connect to the database and run
queries on the metadata log table. An endpoint was created to access
information from those queries and return the data as a json array.
There is the beginning of a testing framework that connects to the db
and checks that the seeded metadata is there.

Lots more to do but you gotta start somewhere.