Add logging checking files are moved correctly
We've been seeing errors in the performance manager where a static
schedule has been ingested twice. Logs from the ingestion pipeline also
show instances where a schedule has been processed twice. The current
theory is that our s3 move files function is not correctly deleting the
incoming static schedule the first time, and its getting processed
twice.

To confirm this, `Converter.convert` has been modified to return a list
of all the files it ingested. In the main method, we now take that list
and intersect it with any files that are still in the incoming bucket.
Any files that haven't been properly moved will be logged.

Additional changes in this commit:
* some comment improvements i made while i was working out how
  everything in the gtfs_rt threading work
* process logger in the `verify_environment` function that logs all the
  env vars we're using.