Write Feed Info Schedule File Last (#145)
* Write Feed Info Schedule File Last

There is a race condition between Ingestion and Performance Manager when
writing the parquet files from a gtfs static schedule. The Performance
Manager understands there to be a new schedule to read when the Feed
Info file is written to the metadata log. In processing this schedule,
it will build other URLs for other files in the schedule. If the Feed
Info file is written before these other files, it could trigger
processing of files that have yet to be written.

To fix this, we want to write the Feed Info table last, so that the
Performance Manager will only attempt to read a static schedule when all
of the files associated with it have been written.

* change some of the structure of the GtfsConverter class to skip
  writing the feed info file while looping through zip contents and then
  write the feed info file. only write metadata for feed info file.
* generate the timestamp key for static schedules from the time that the
  feed_info file was written. (i checked a months worth and these files
  are generally written an hour or two before delta processes them).
* remove last_modified timestamp from return signature of
  `get_zip_buffer` function
* update tests to accommodate this, including monkey patching the write
  parquet and move s3 objects functions.