feat: CSV Flatfile Pipeline Testing (#154)
This change provides a new testing method test_whole_table. This method takes a CSV representation of Vehicle Position records, converts it to a Parquet file on the fly and runs it through process_gtfs_rt_files to populate GTFS-RT RDS tables.

After Processing, a query selects records from the RDS and compares them against a CSV results file to confirm that there have been no fundamental changes to the Performance Manager processing pipelines.

The input CSV file consists of mostly whole route trips, 3 in each direction for each MBTA Rail line.

This PR required the updating of GTFS and GTFS-RT test files in the SPRINGBOARD test_files folder of the repository to match that date of records from the flat file (May 8th 2023). In an effort to reduce the repository size, all non rail data has also been stripped from the SPRINGBOARD parquet files of RT_TRIP_UPDATES and RT_VEHICLE_POSITIONS.

This Branch was developed with a commit from prior to the inclusion of the PR that removed hash columns from the application. This was done to provide a check against the validity of the changes introduced by that PR.

Asana Task: https://app.asana.com/0/1204931901750655/1205084207879142