Add function for generating flat pyarrow table
For some external users, we're going to provide flat parquet files via a
data portal that can be used to run analysis on the data we've
synthesized in the rail performance manager. Create a function that
generates the pyarrow table with the appropriate data types for a given
service date. This function will be run on a daily cadence, with the
resulting table being written to a parquet file in an accessible s3
bucket.

https://app.asana.com/0/1204931901750655/1204907226653447/f