Create a migration config for dev, staging, & prod (#164)
* Create a migration config for dev, staging, & prod

We want to be able to specify migrations on an environmental basis.
Create a versions folder for staging (thats currently a copy of prod)
and move the generic performance_manager version folder to
performance_manager_dev.

Additionally, add the alembic_db_name variable to the environment
validation function in the performance manager pipeline and remove the
fallback value, ensuring that it must be specified by the container
running it.

Update CI to always use the dev migrations. The staging and prod
migrations should be compressions of the dev ones, so dev should
always be the latest we're developing against.