Update Environment Validation method
I wanted to start logging environment variables for performance manager
the same way we do for ingestion. When i made that update, pylint
correctly yelled at me for duplicating code, so I moved all validation
into its own function in its own file. In doing so I made it a bit more
generic, so we now specify required parameters when making the call and
if we want to check against database parameters. I also added a piece
that allows us to check for optional parameters (like the public archive
bucket) that will be logged iff they exist.

This function is now imported into the pipeline files for both
performance manager and ingestion. It should be used in other, future,
pipelines as well.