Proof of Concept for Flashlight
Flashlight is a small elixir library intended to be used to read small
bits of archived lamp data by various other applications at ctd. For a
proof of concept, I worked out how to extract, join, and filter busloc
and vehcile position data for a given run id.

This commit contains the initial work for setting up the library with
dependencies and running the methods in iex. To call it a work in
progress would be a compliment.