FIX: Use Trip Updates to help compute start time
When processing trips, we need to compute a start time for added trips.
We've coded doing that based on the minimal vp move and vp stop
timestamps for a given trip. However, this flow triggers a failure if
there are only Trip Update records for this event. In those cases, use
the minimal tu stop timestamp as well.