Clean Up Static Data if Load Fails
We process static schedule files in series, extracting data,
transforming it, and loading it into the rail performance manager
database. If there is a failure in processing one of them, data from
earlier files will be inserted, the error is raised, and data from later
files will be missing. Rather than leave some of this information in the
database, catch these errors and delete incomplete records before
re-raising the error.