change trip_id static fields from 128 to 512 limit
The strings in some trip_ids are a but longer than the 128 limit we've
been using. Quadruple it so that we don't run into this error again.

Add a new migration script in staging and prod to get us to the
appropriate state.

NOTE: there is a piece related to the `static_service_id_lookup` table
in these migration files resulting from a previous oversight.