Move existing Metadata from RPM to MD
We need existing metadata to be copied from the rail performance manager
rds to the metadata rds. Do so in the ingestion application before
starting the main method.

In the performance manager, don't enter the biz logic of an event loop
until the legacy metadata table has been copied over.