Fix rds writer process and with more typing (#217)
The RDS writer process waits for items to be put on a metadata queue,
pops them off, and then writes that file to the metadata. In the past,
the queue was typed with any, as it was being populated with an obscure
pyarrow object we couldn't deduce properly.

With the recent changes to how we write parquet files, these changed to
strings. This caused the process to fail since it was attempting to pull
a `path` attribute out a string, causing an uncaught exception.

I've updated the process to expect a string from the queue and put in
typing hints that should prevent an error like this in the future.