FIX: Allow empty calendar_dates GTFS static table (#258)
When performance_manager is importing a GTFS static schedule, it expects all ingested tables to contain data. Per the GTFS Spec, the calendar_dates table is not required to contain data.

Out ingestion application manifests this as not writing any parquet file to S3 for an empty calendar_dates static file. file_list_from_s3 then returns an empty list when the CALENDAR_DATES partition is listed for a specific static schedule.

To fix this, we will catch a KeyError when converting a GTFS static parquet file into a pandas dataframe, and return an empty dateframe with all required columns, if allowed by the allow_empty_dataframe bool flag of the StaticTableColumns dataclass.

Asana Task: https://app.asana.com/0/1205827492903547/1206643238763137