FIX: Reduce resource usage during parquet file creation (#267)
The LAMP staging environment has been displaying excessively high memory usage and crashes during the creation of parquet files for Tableau publishing. The issue was traced back to the write_to_parquet function of our DatabaseManager object.

SQLAlchemy result streaming through use of the yield_per method, was not behaving as intended. Instead of incrementally streaming results from a query, it was materializing the entire result stream in memory before iterating through the results. Updating the write_to_parquet function to use the execute_options parameter has resolved this issue and results are now correctly streamed incrementally and memory usage during streaming is contained to a single batch_size amount.

Additionally, this change adds helper functions to our DatabaseManager objects to ENABLE and DISABLE the rt_trips_update_branch_trunk TRIGGER of our vehicle_trips table. This trigger was causing excessively long query team for an UPDATE operation against the vehicle_trips table and is generally not required. This is a bit of a band-aid fix for now, that can be fixed with architectural RDS and/or application changes in the future.

Asana Task: https://app.asana.com/0/1205827492903547/1206703823960612