FEAT: Skip upload of LAMP_ALL_RT_fields.parquet file when no new records (#269)
Currently, the LAMP_ALL_RT_fields.parquet file is uploaded to S3 on every event loop.

The change adds a check between the records >= max_start_date in the database and in the existing LAMP_ALL_RT_fields.parquet file. If the counts are equivalent, the parquet file is not re-uploaded to S3.

Also includes all dependabot updates for the week.