FIX: Remove vestigial dot pylintrc
Our pylint configuration is in `pyproject.toml`. We had a `.pylintrc`
file from before times. It didn't matter when we had a python_src
directory, but now that everything is at the top level it causes
problems when running pylint via poetry locally. (It still worked
correctly on github because we don't copy the .pylintrc file into our
testing environments).

Remove that file to avoid conflicts in local development.