Order the daily tables by stop time.
In communication with Transit Matters, they've asked that we sort our
daily rail performance manager parquet files by stop time. To do that,
order our flat file query by a coalesced timestamp from the multiple
move and stop timestamps on a vehicle event.

This is being done in the same PR for publishing prod data, this tasks
asana ticket is:
https://app.asana.com/0/1205827492903547/1206935232799271/f