FIX Correct Parquet Generation Assertion
When writing parquet file, we “partition” files by creating Tables that
have have the same value in all of the columns we're plan on
partitioning by before writing. To ensure we processed data correctly
before writing, the `write_parquet_file` has an assertion that the list
of unique values in partition columns is exactly one. However, if the
Table that is being passed in is empty, that list of unique values will
have zero, and the assertion fill throw an exception.

* Add an additional `if` gate that checks that the gtfs table is not
  empty before writing it to s3.
* Add a message in the assertion that will allow us to debug this
  problem easier going forward.