Add Versioning to Docker Images (#383)
Add a `__version__.py`file to the lamp_py module. This file is left with
a default value for local development. In our Dockerfile, overwrite this
file with a passed in version argument, allowing the version to be set
inside of the docker image. Update the deploy base github action to
generate a unique version value for each deployment based on the latest
version tag and current commit hash. Lastly, log the version when
validating the environment.