Update Glides Parquet Files (#397)
* Update Glides Parquet Files

In the initial deployment of glides ingestion, location fields that
include the gtfsId and todsId fields incorrectly used ID instead of Id.
Correct this in the schemas for each of our record types and create a
migration that will modify existing data to the new schema so that old
and new datasets can be joined correctly.

While investigating, I also noticed that there were duplicate records
in our files stemming from restarting the application and re-ingesting
data from the past 24 hours. In the migration and in the
GlidesConverter class, deduplicate the datasets based on the 'id'
column.