SQUASH ON MERGE
Refactor how we choose files. Create a dataframe that contains all of
the source files from gtfs rt and transit master and generate service
dates for each one. Then it compares last modified datetimes with a
latest file to find service dates that need to be updated. Then it pulls
out all of the files related to service dates that need to be updated.