FIX: GTFS Compressed Schedule Upload (#447)
The S3 sync operation for GTFS compressed parquet files was calling the "GTFSArchive.parquet_path" method to create object upload paths. This method automatically appends ".parquet" to any file passed to it. However the GTFS_ARCHIVE.db.gz file is not supposed to have a ".parquet" appended.

This change stops using the "GTFSArchive.parquet_path" method and falls back to a simple path join to create the S3 object upload path. The matches the behavior of the current PROD environment which has not seen these errors.